refactor: split codes to utils.js, template.js
