+++ index.html
@@ -136,11 +136,13 @@
     <div>
         <script type="module" src="./route.js"></script>
         <script type="module" src="./app.js"></script>
-        <script type="module" src="./Login/login.js"></script>
-        <script type="module" src="./Home/home.js"></script>
-        <script type="module" src="./Chat/chat.js"></script>
-        <script type="module" src="./Game/game.js"></script>
-        <script type="module" src="./Rank/rank.js"></script>
+        <script type="module" src="./Login/loginTemplate.js"></script>
+        <script type="module" src="./Home/homeTemplate.js"></script>
+        <script type="module" src="./Home/homeUtils.js"></script>
+        <script type="module" src="./Chat/chatTemplate.js"></script>
+        <script type="module" src="./Chat/chatUtils.js"></script>
+        <script type="module" src="./Game/gameTemplate.js"></script>
+        <script type="module" src="./Rank/rankTemplate.js"></script>
     </div>
 </body>
 </html>
